Add guild ID to message.
A bot/client can be in multiple guilds simultaneously. As a result,
it is useful to be able to distinguish from which guild a message
was originated, without having to either query Discord on each
message (to ask which guild owns a channel ID) or store that
information ourselves in our apps.

Signed-off-by: Rick Elrod <rick@elrod.me>