Change 'cabal configure' to the equivalent 'cabal build --dry-run'
After cabal 3.8.1.0, 'cabal configure' only performs things related to
cabal.project.local, and does not generate a build plan any more. The old
behaviour can still be achieved through 'cabal build --dry-run', which
is also present in older cabal versions. Having asked on #hackage, it
seems this is an intentional change over the course of several cabal PRs,
so it's best to keep it future-proof.